Expe tool: exec N instances working
Many instances can be executed thanks to Execo.
Instances already done in the past will not be recomputed :).

Variables can now be either:
  - simple. one value, accessed via ${VAR}
  - lists. accessed via @n{VAR}, where n is
    the element to access.

Remote execution is implemented but has not been tested yet.

TODO: pre and post commands of the exec1 instance
TODO: pre and post commands of the execN instances
TODO: hide error messages in the execN script?