Exec: var decl order. ExecN: post/pre cmds
The execution scripts now handles properly the
declaration order of variables. To do so, the scripts
find a declaration order and checks that the dependency
graph between variables is acyclic.

Post and pre commands can now be used in the
ExecN instances script.

The test_energy input file of these scripts have been
modified to test those behaviours