Expe tool: bat command is better managed now
This commit fixes issue #2, thanks to pfdutot's patch.

Previously, the user-given was directly parsed via argparse.
This limited the user-given command to be a direct call to
Batsim. Furthermore, the script was directly aborted on
parsing error, which did not explicit the abort reason well.

Now, prefixed Batsim calls should be handled.
Furthermore, whenever argparse fails in parsing the command,
the error outputting should be way clearer.