Exec: only recompute post-cmds of instances done
This commit adds the --post_only option of exec1 script.
This option only computes the post_commands of the instance.

This commit also adds the --recompute_instances_post_commands of
the execN script. This option allows to only recompute the
post_commands of all the instances which were marked as done
before running execN.

-> Now you can avoid recomputing long instances while modifying
   your post_commands!