Bugfix: date string buffer was too short.
The date is written as a simple floating-point number in the protocol.
However, the size of this date buffer was 16, which caused simulations longer
than 1e8 to fail. This happened to me when I tried to simulate HPC2N.

This buffer is now SIZE=32-character long.
Furthermore, the simulation is now stopped whenever a (SIZE-1)-character long
date is encountered, as it is susceptible to lose information, which may cause
infinite loops.