ExecN: unique instance IDs are now generated
These instance IDs can be used in the YAML scripts,
to generate filenames for example. This would allow
to bypass the maximum filename limitation.

These instance IDs are also logged in a CSV file,
located in base_output_dir/instances/instances_info.csv.
This would allow simple and generic merge of information
in the postprocessing.