[misc] version from CMake (from git if possible)
Version is no longer hardcoded in the source code.
The source code simply reads its BATSIM_VERSION macro.

BATSIM_VERSION is set by CMake.
(if Batsim is compiled without it, it sets version as vUNKNOWN_PLEASE_COMPILE_VIA_CMAKE)

Version is retrieved from git-describe.
If git is not available or if Batsim is built from a tarball, a default version is used.
This default version is hardcoded in CMake.