[code] dynamic_submit: sooner memory update, log++
On SUBMIT_PROFILE, the code to add the profile into memory is now done
in the protocol reader rather than in the orchestrator.

This allows to manage correctly a SUBMIT_JOB event after a
SUBMIT_PROFILE one in the same message.

Furthermore, the user should now have warnings with clear text when
doing error-prone actions with dynamic submissions.