[code] workload: prevent direct construction
Two constructors existed either for static or dynamic workloads, but
this semantic was unclear.

Two static methods (new_static_workload and new_dynamic_workload) now
exist in the Workload class for this purpose, and other constructors
should not be accessible from outside the class.

Furthermore, whether the workloads are static or not is now stored in
the class. This allows to know how many static workloads are in the
simulation (which will be required for a proper termination).