[code] cleaner termination
The previous code to terminate the simulation was quite complex and
misleading.

The current one does essentially the same but stores another state to
make things simpler. An is_simulation_finished() function has been added
for this purpose. The main loop is now simply to wait until the
scheduler awknowledges the end of the simulation.