[test] disable kill.*.*after0 tests
Related to https://gitlab.inria.fr/batsim/batsim/issues/37.

We agreed on disabling these tests for the moments.
Ideally, the tests should still exist and be marked as skipped
(or marked as expected to fail), but cmake does not allow this so
they are just not generated for the moment.
pytest would allow this to be clearer.